Fix poison for address

+++ asm/src/code/asm.c
@@ -247,7 +247,7 @@ static enum ASM_CODES asmMakeInstr(cpuInstruction_t *dest, const char *cmd, cons
     {
         dest->opcode.cmd = CMD_JMP;
 
-        cpuData_t argVal = {.address = SIZE_MAX};
+        cpuData_t argVal = {.address = LLONG_MAX};
         if (1 == sscanf(arg, "%%%lld", &argVal.address)) { ; }
         else if (SIZE_MAX != (argVal.address = asmFindLabel(arg, labels))) { ; }
         else { ; }
